+++ lib/screens/sign_up/components/sign_up_form.dart
@@ -1,0 +1,160 @@
+import 'package:ecommerce/components/custom_surfix_icon.dart';
+import 'package:ecommerce/components/default_button.dart';
+import 'package:ecommerce/components/form_error.dart';
+import 'package:flutter/material.dart';
+
+import '../../../constants.dart';
+import '../../../size_config.dart';
+
+class SignUpForm extends StatefulWidget {
+  @override
+  _SignUpFormState createState() => _SignUpFormState();
+}
+
+class _SignUpFormState extends State<SignUpForm> {
+  String email;
+  String password;
+  String confirm_password;
+  final List<String> errors = [];
+  final _formKey = GlobalKey<FormState>();
+  bool remember = false;
+
+
+  void addError({String error}) {
+    if(!errors.contains(error))
+      setState(() {
+        errors.add(error);
+      });
+  }
+  void removeError({String error}) {
+    if(!errors.contains(error))
+      setState(() {
+        errors.remove(error);
+      });
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Form(
+      key: _formKey,
+      child: Column(
+        children: [
+          buildEmailFormField(),
+          SizedBox(height: getProportionateScreenHeight(30),),
+          buildPassFormField(),
+          SizedBox(height: getProportionateScreenHeight(30),),
+          buildConfirmPassFormField(),
+          SizedBox(height: getProportionateScreenHeight(30),),
+          FormError(errors: errors),
+          DefaultButton(text: "Continue",
+            press: () {
+              if(_formKey.currentState.validate()){
+                _formKey.currentState.save();
+                //Go to complete profile
+              }
+            },)
+        ],
+      ),
+    );
+  }
+
+  TextFormField buildPassFormField() {
+    return TextFormField(
+      obscureText: true,
+      onSaved: (newValue) => password = newValue,
+      onChanged: (value) {
+        if (value.isNotEmpty ){
+          removeError(error: kPassNullError );
+        }else if(value.length >= 8){
+          removeError(error: kShortNullError);
+        }
+        password = value;
+
+      },
+      validator: (value){
+        if (value.isEmpty ){
+          addError(error: kPassNullError );
+          return "";
+        }else if(value.length < 8){
+          addError(error: kShortNullError);
+          return "";
+        }
+        return null;
+      },
+      decoration: InputDecoration(
+        labelText: "Password",
+        hintText: "Input your Password",
+        floatingLabelBehavior: FloatingLabelBehavior.always,
+        suffixIcon: CustomSuffixIcon(
+          svgIcon: "assets/icons/Lock.svg",
+        ),
+      ),
+    );
+  }
+
+  TextFormField buildConfirmPassFormField() {
+    return TextFormField(
+      obscureText: true,
+      onChanged: (value) {
+        if (value.isNotEmpty ){
+          removeError(error: kPassNullError );
+        }else if(value.isNotEmpty && password == confirm_password){
+          removeError(error: kMatchNullError);
+        }
+        confirm_password = value;
+
+      },
+      validator: (value){
+        if (value.isEmpty ){
+          addError(error: kPassNullError );
+          return "";
+        }else if((password != value)){
+          addError(error: kMatchNullError);
+          return "";
+        }
+        return null;
+      },
+      decoration: InputDecoration(
+        labelText: "Confirm Password",
+        hintText: "Confirm your Password",
+        floatingLabelBehavior: FloatingLabelBehavior.always,
+        suffixIcon: CustomSuffixIcon(
+          svgIcon: "assets/icons/Lock.svg",
+        ),
+      ),
+    );
+  }
+
+  TextFormField buildEmailFormField() {
+    return TextFormField(
+      keyboardType: TextInputType.emailAddress,
+      onSaved:  (newValue) => email = newValue,
+      onChanged: (value) {
+        if (value.isNotEmpty){
+          removeError(error: kEmailNullError);
+        }else if(emailValidatorRegExp.hasMatch(value)){
+          removeError(error: kInvalidEmailError);
+        }
+        return null;
+      },
+      validator: (value){
+        if (value.isEmpty){
+          addError(error: kEmailNullError);
+        }else if(!emailValidatorRegExp.hasMatch(value)){
+          addError(error: kInvalidEmailError);
+          return "";
+        }
+        return null;
+      },
+      decoration:
+      InputDecoration(
+        labelText: "Email",
+        hintText: "Input your Email",
+        floatingLabelBehavior: FloatingLabelBehavior.always,
+        suffixIcon: CustomSuffixIcon(
+          svgIcon: "assets/icons/Mail.svg",
+        ),
+      ),
+    );
+  }
+}
